optionlist2

+++ src/main/java/com/ssghot/ssg/optionList/repository/IStockRepository.java
@@ -3,5 +3,5 @@ package com.ssghot.ssg.optionList.repository;
 import com.ssghot.ssg.optionList.domain.Stock;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IStockRepository extends JpaRepository<Stock, Integer> {
+public interface IStockRepository extends JpaRepository<Stock, Long> {
 }
